+++ lib/commons_lang.dart
@@ -21,3 +21,5 @@ library;
 export 'src/bool_utils.dart';
 export 'src/text/str_lookup.dart';
 export 'src/text/str_matcher.dart';
+export 'src/text/str_builder.dart';
+export 'src/text/str_substitutor.dart';